fixed a problem with reloading the page on registerforstudents when a validation error occurs, the page cann't reload drop downs.
need to fix situation where you change a major it registers the new major and doesn't change the old registration participation.  happened when there was 2 reg fields (one from student's major) and another from changed participation.  then changing the majors to match but not registering for the one from the student's major.

git-tfs-id: [https://tfs.caes.ucdavis.edu:8088/tfs]$/Commencement/Trunk/Commencement;C8951