Nicer error page
If no student found for the current term, go to nice error page.
This can happen once a student has graduated and is not in banner for the current term.
Can also happen if the student doesn't have the minimum units.
If they have the units, but have already walked, eventually they would get a different error page.